requirements.txt
aiogram==3.6.0
telethon==1.40.0
python-dotenv==1.0.1
APScheduler==3.10.4
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.20
